+++ evm/test/CatalystVault/Set/SetVaultFee.t.sol
@@ -52,7 +52,7 @@ abstract contract TestSetVaultFee is Test, AVaultInterfaces {
     function test_set_fee_not_too_large() external virtual {
         address[] memory vaults = getTestConfig();
 
-        uint256 vaultFee = 10**18;
+        uint64 vaultFee = 10**18;
 
         for (uint256 i = 0; i < vaults.length; ++i) {
             address vault = vaults[i];
@@ -67,7 +67,7 @@ abstract contract TestSetVaultFee is Test, AVaultInterfaces {
     function test_error_set_fee_too_large() external virtual {
         address[] memory vaults = getTestConfig();
 
-        uint256 vaultFee = 10**18 + 1;
+        uint64 vaultFee = 10**18 + 1;
 
         for (uint256 i = 0; i < vaults.length; ++i) {
             address vault = vaults[i];
